ColorMode: export ColorMode function
