+++ src/ColorMode/SliderColorMode.css
