+++ src/Cookie-Alert/assets/style.js
@@ -1,4 +1,4 @@
-import {darkModeSelecting} from '../../Navbar/Navbar';
+import {darkModeSelecting} from '../../ColorMode/ColorMode';
 
 export const cookieBannerbackgroundColor = () => {
   if (darkModeSelecting() === true) {
